Add trusting game implementation
Add an implementation of the CardGame trait that communicates exclusively in
plaintext. This allows users of the library to build and test with
easily-debuggable inputs and outputs. The intent is that the local instance of
the trusting CardGame implementation will apply the same constraints on inputs
and outputs that the crptographic version (still to be written) will.